requirements.txt
aws-cdk-lib==2.201.0
constructs>=10.0.0,<11.0.0
langchain==0.1.0
faiss-cpu==1.7.4
boto3==1.34.0
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
requests==2.31.0
